refactor(core): move skill discovery report lookup into skill manager
